+++ packages/components/src/Button/GenericButton/GenericButton.tsx
@@ -8,8 +8,8 @@ import React, {
   FocusEvent,
 } from "react"
 import classnames from "classnames"
-import { Badge, BadgeAnimated } from "@kaizen/draft-badge"
-import { LoadingSpinner } from "@kaizen/loading-spinner"
+import { Badge, BadgeAnimated } from "~components/Badge"
+import { LoadingSpinner } from "~components/Loading"
 import styles from "./GenericButton.module.scss"
 
 export type CustomButtonProps = {