+++ packages/components/src/Filter/FilterButton/FilterButtonRemovable/FilterButtonRemovable.tsx
@@ -1,10 +1,10 @@
 import React, { forwardRef } from "react"
-import { Tooltip } from "@kaizen/draft-tooltip"
 import { ButtonGroup, ButtonGroupProps } from "~components/ButtonGroup"
+import { FilterTriggerRef } from "~components/Filter/Filter"
 import { ClearIcon } from "~components/Icon"
+import { Tooltip } from "~components/Tooltip"
 import { DataAttributes } from "~types/DataAttributes"
 import { isRefObject } from "~utils/isRefObject"
-import { FilterTriggerRef } from "../../Filter"
 import { FilterButton, FilterButtonProps } from "../FilterButton"
 import {
   FilterButtonBase,