+++ packages/components/src/Illustration/Scene/index.ts
@@ -1,1 +1,2 @@
+export * from "./BrandMomentCaptureIntro"
 export * from "./Scene"