+++ packages/components/src/Loading/LoadingGraphic/_docs/LoadingGraphic.stories.tsx
@@ -1,11 +1,8 @@
 import React from "react"
 import { Meta, StoryObj } from "@storybook/react"
-import {
-  BrandMomentPositiveOutro,
-  Informative,
-} from "@kaizen/draft-illustration"
 import { Avatar } from "~components/Avatar"
 import { ConfigureIcon } from "~components/Icon"
+import { BrandMomentPositiveOutro, Informative } from "~components/Illustration"
 import { StickerSheet } from "~storybook/components/StickerSheet"
 import { LoadingGraphic } from "../index"