+++ packages/components/src/__future__/Select/utils/getDisabledKeysFromItems.ts
@@ -1,3 +1,4 @@
+import { Key } from "@react-types/shared"
 import { SelectItem, SelectOption } from "../types"
 import { isSelectOptionGroup } from "./isSelectOptionGroup"
 
@@ -5,8 +6,8 @@ export const getDisabledKeysFromItems = <
   Option extends SelectOption = SelectOption,
 >(
   items: Array<SelectItem<Option>>
-): React.Key[] =>
-  items.reduce((acc: React.Key[], item) => {
+): Key[] =>
+  items.reduce((acc: Key[], item) => {
     if (isSelectOptionGroup(item)) {
       const keys = Array.from(item.options)
         .filter(groupItem => groupItem.disabled)