+++ storybook/preview.tsx
@@ -1,17 +1,12 @@
+import "./styles/styles.scss"
 import "./styles/tailwind.scss"
+
 import React, { useEffect } from "react"
 import { Preview } from "@storybook/react"
 import isChromatic from "chromatic"
 import { KaizenProvider } from "~components/KaizenProvider"
 import { backgrounds } from "./backgrounds"
 import { DefaultDocsContainer } from "./components/DocsContainer"
-
-import "highlight.js/styles/a11y-light.css"
-
-// Standard base stylesheet used across Culture Amp products
-// See: https://github.com/necolas/normalize.css/
-import "normalize.css"
-import "@kaizen/component-library/styles/fonts.scss"
 import { globalA11yRules } from "./global-a11y-rules"
 
 const IS_CHROMATIC = isChromatic()